chore: move theme demo
