+++ src/modules/wallet/components/walletVisual/demo.js
@@ -2,7 +2,7 @@ import React, { useState } from 'react';
 import Waypoint from 'react-waypoint';
 import { extractAddressFromPassphrase } from '@wallet/utils/account';
 import { generatePassphraseFromSeed } from 'src/modules/wallet/utils/passphrase';
-import DemoRenderer from '@basics/demoRenderer';
+import DemoRenderer from 'src/theme/demo/demoRenderer';
 import WalletVisual from '.';
 
 /**
